[audio] Use float audio data in AudioData
Most audio libraries seem to work with a float32 sample format.
Another advantage is it doesn't clip & distort during processing.

Convert to uint8 sample format only when requesting legacy "visdata"
arrays, as long as effects still expect this.

+++ avs/vis_avs/avs.cpp
@@ -79,8 +79,8 @@ bool avs_render_frame(AVS_Handle avs,
 
 AVS_API
 int32_t avs_audio_set(AVS_Handle avs,
-                      const int16_t* left,
-                      const int16_t* right,
+                      const float* left,
+                      const float* right,
                       size_t audio_length,
                       size_t samples_per_second) {
     return -1;
